WIP: Dashboard Screen - Adding Coil plus Pokemon Details
Coil added for images plus changes in how Polemons are being displayed, added call to retrieve pokemon details

+++ app/src/main/java/com/carlos/pokedex/core/network/Mapper.kt
@@ -22,7 +22,7 @@ fun PokemonDto.toDomain(): PokemonDetails {
     return PokemonDetails(
         name = name.replaceFirstChar { it.uppercase() },
         imageUrl = sprites.frontDefault,
-        height = height,
-        weight = weight
+        height = height.toInt(),
+        weight = weight.toInt()
     )
 }
